edit working

+++ src/store/actions/projectActions.js
@@ -13,7 +13,12 @@
  */
 export const createProject = (project) => {
 
+
+
     return (dispatch, getState, { getFirebase }) => {
+
+
+
         //Make async call to database
         const firestore = getFirebase().firestore();
         const profile = getState().firebase.profile;
@@ -36,17 +41,56 @@ export const createProject = (project) => {
 
 export const deleteProject = (project) => {
 
+
     return (dispatch, getState, { getFirebase }) => {
+
         //Make async call to database
         const firestore = getFirebase().firestore();
-        firestore.collection('projects')
-            .doc(project.dogname)
-            .delete().
-            then(() => {
-            // Dispatch goes to reducers -> projectReducser
-            dispatch({ type: 'DELETE_PROJECT', project });
-        }).catch((err) => {
-            dispatch({ type: 'DELETE_PROJECT_ERROR', err })
-        });
+        const profile = getState().firebase.profile;
+        const authId = getState().firebase.auth.uid;
+
+        if (authId == project.authorId){
+            firestore.collection('projects')
+                .doc(project.dogname)
+                .delete().
+                then(() => {
+                // Dispatch goes to reducers -> projectReducser
+                dispatch({ type: 'DELETE_PROJECT', project });
+            }).catch((err) => {
+                dispatch({ type: 'DELETE_PROJECT_ERROR', err })
+            });
+        } else {
+            console.log("You are not current user!")
+        }
+
+
+
+    }
+};
+
+
+export const editProject = (project) => {
+
+
+    return (dispatch, getState, { getFirebase }) => {
+
+        //Make async call to database
+        const firestore = getFirebase().firestore();
+        const profile = getState().firebase.profile;
+        const authId = getState().firebase.auth.uid;
+
+        firestore.collection('projects').doc(project.dogname).update({
+            ...project,
+            authorFirstName: profile.firstName,
+            authorLastName: profile.lastName,
+            authorId: authId,
+            createdAt: new Date(),
+            }).then(() => {
+                // Dispatch goes to reducers -> projectReducer
+                dispatch({ type: 'EDIT_PROJECT', project });
+            }).catch((err) => {
+                dispatch({ type: 'EDIT_PROJECT_ERROR', err })
+            });
+
     }
 };
